Added Files by Upload
This is the first and most simple commit. This project doesn't have any unit tests or complexities. An algorithim has to be added to perform the calculation of bills and coins but the purpose of this commit is just to complete the first exercise. This exercise will be later revisited to be improved.